buzzfeed_forum/requirements.txt
Django==1.8.6
django-filter==0.11.0
djangorestframework==3.3.1
Markdown==2.6.3
